+++ CoreLab/labs/labs/LabThree.java
@@ -13,8 +13,8 @@ public class LabThree {
 	public static void main(String args[]) {
 		problemOne();
 		problemTwo();
-		problemThree(0.03);
-		problemFour(0.08);
+		problemThree(0.03f);
+		problemFour(0.08f);
 	}
 	
 	
@@ -42,7 +42,7 @@ public class LabThree {
 	
 	// a histogram of the Paolina image with the top and bottom
 	// cutoffPercent of points removed
-	public static void problemThree(double cutoffPercent) {
+	public static void problemThree(float cutoffPercent) {
 		File outputFile1 = new File(WORK_DIR+"problem-three-Paolina.png");
 		File outputFile2 = new File(WORK_DIR+"problem-three-histogram.png");
 		
@@ -52,7 +52,7 @@ public class LabThree {
 		ImageProcessor.writeImage(proc.cutoffHisto(cutoffPercent), outputFile2);
 	}
 	
-	public static void problemFour(double cutoffPercent) {
+	public static void problemFour(float cutoffPercent) {
 		File myFile = new File(WORK_DIR+"haswell-e-die-shot.jpg");
 		File outputFile1 = new File(WORK_DIR+"problem-four-original.png");
 		File outputFile2 = new File(WORK_DIR+"problem-four-original-histogram.png");
